third commit

+++ app/Http/Controllers/BannerController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers;
 use App\Models\Banner;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Storage;
+use Illuminate\Support\Facades\Gate;
 
 class BannerController extends Controller
 {
@@ -25,6 +26,9 @@ class BannerController extends Controller
      */
     public function store(Request $request)
     {
+        if (!Gate::forUser($request->user())->allows('manager-action')) {
+            return response()->json(["message"=>"Bạn không phải là manager, bạn không có quyền này ?"],403);
+        }
         $banner = new Banner();
         $banner->image = $request->file("banner_image")->store("images/banners");
         $banner->short_description = $request->short_description;
@@ -46,6 +50,9 @@ class BannerController extends Controller
      */
     public function update(Request $request, Banner $banner)
     {
+        if (!Gate::forUser($request->user())->allows('manager-action')) {
+            return response()->json(["message"=>"Bạn không phải là manager, bạn không có quyền này ?"],403);
+        }
         if($request->has("banner_image")){
             Storage::delete($banner->image);
             $banner->image = $request->file("banner_image")->store("images/banners");
@@ -59,8 +66,11 @@ class BannerController extends Controller
     /**
      * Remove the specified resource from storage.
      */
-    public function destroy(Banner $banner)
+    public function destroy(Request $request,Banner $banner)
     {
+        if (!Gate::forUser($request->user())->allows('manager-action')) {
+            return response()->json(["message"=>"Bạn không phải là manager, bạn không có quyền này ?"],403);
+        }
         Storage::delete($banner->image);
         $banner->delete();
         return response()->json(["message"=>"delete ok"],200);
