+++ app/Http/Controllers/CategoryAttributeValueController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers;
 
 use App\Models\CategoryAttributeValue;
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 
 class CategoryAttributeValueController extends Controller
 {
@@ -25,6 +26,9 @@ class CategoryAttributeValueController extends Controller
      */
     public function store(Request $request)
     {
+        if (!Gate::forUser($request->user())->allows('manager-action')) {
+            return response()->json(["message"=>"Bạn không phải là manager, bạn không có quyền này ?"],403);
+        }
         return CategoryAttributeValue::create([
             "value" => $request->input("value"),
             "category_attribute_id" => $request->input("category_attribute_id")
@@ -44,6 +48,9 @@ class CategoryAttributeValueController extends Controller
      */
     public function update(Request $request, CategoryAttributeValue $categoryAttributeValue)
     {
+        if (!Gate::forUser($request->user())->allows('manager-action')) {
+            return response()->json(["message"=>"Bạn không phải là manager, bạn không có quyền này ?"],403);
+        }
         $categoryAttributeValue->value = $request->input("value");
         $categoryAttributeValue->save();
         return response()->json(["message" => "Cập nhật thành công"], 200);
@@ -52,8 +59,11 @@ class CategoryAttributeValueController extends Controller
     /**
      * Remove the specified resource from storage.
      */
-    public function destroy(CategoryAttributeValue $categoryAttributeValue)
+    public function destroy(Request $request,CategoryAttributeValue $categoryAttributeValue)
     {
+        if (!Gate::forUser($request->user())->allows('manager-action')) {
+            return response()->json(["message"=>"Bạn không phải là manager, bạn không có quyền này ?"],403);
+        }
         $categoryAttributeValue->delete();
         return response()->json(['message' => "OK"], 200);
     }
